% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_scp.R
\name{scp}
\alias{scp}
\title{scp}
\usage{
scp(jShellServer, fromFile, toFile, passRemote, portRemote = "22",
  params = "-q", compression = FALSE,
  bandwithLimit = default.bandwithLimit, RUN = TRUE)
}
\description{
\code{scp} Uses system scp through ssh connection to move a file from connected server to a different server.
\cr This function is exactly the same than \code{scpToOther}.
}
\details{
Uses system scp through ssh connection to move a file from connected server to a different server.
\cr To move a whole folder, and all recursive subfolders, use \code{\link{scpToOther}} or \code{\link{scpFromOther}} with the following parameters: params="-q -r"
}
\seealso{
\code{\link{scpToLocal}} \code{\link{scpFromLocal}}
\cr \code{\link{scpToOther}} \code{\link{scpFromOther}}
\cr \code{\link{scp}}
}

