% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_scp.R
\name{scpFromLocal}
\alias{scpFromLocal}
\title{scpFromLocal}
\usage{
scpFromLocal(jShellServer, fromLocalFile, toRemoteFile)
}
\description{
\code{scpFromLocal} It does not uses system scp. It uses java process. It moves a file from local server to connected server.
}
\details{
It does not uses system scp. It uses java process. It moves a file from local server to connected server.
\cr To move a whole folder, and all recursive subfolders, use \code{\link{scpToOther}} or \code{\link{scpFromOther}} with the following parameters: params="-q -r"
}
\examples{
\dontrun{
con = ssh.shell()
ssh.open(con, connData$server_sas_ip, connData$server_sas_port, connData$server_sas_user, connData$server_sas_pass)
write.table(iris, file="iris.csv", append=FALSE, quote=FALSE, sep=";", row.names=FALSE)

scpFromLocal(con, "iris.csv", paste0(connData$server_sas_tables,"iris.csv"))

ssh.close(con)
}
}
\seealso{
\code{\link{scpToLocal}} \code{\link{scpFromLocal}}
\cr \code{\link{scpToOther}} \code{\link{scpFromOther}}
\cr \code{\link{scp}}
}

