% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_scp.R
\name{scpToLocal}
\alias{scpToLocal}
\title{scpToLocal}
\usage{
scpToLocal(jShellServer, fromRemoteFile, toLocalFile)
}
\description{
\code{scpToLocal} It does not uses system scp. It uses java process. It moves a file from connected server to local server.
}
\details{
It does not uses system scp. It uses java process. It moves a file from connected server to local server.
\cr To move a whole folder, and all recursive subfolders, use \code{\link{scpToOther}} or \code{\link{scpFromOther}} with the following parameters: params="-q -r"
}
\seealso{
\code{\link{scpToLocal}} \code{\link{scpFromLocal}}
\cr \code{\link{scpToOther}} \code{\link{scpFromOther}}
\cr \code{\link{scp}}
}

