% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_base.R
\name{ssh.open}
\alias{ssh.open}
\title{ssh.open}
\usage{
ssh.open(jShellServer, server, port, user, pass, bash_prompt = "",
  privateKey = "", strictHostKeyChecking = "", showConfig = FALSE)
}
\examples{
\dontrun{
con = ssh.shell()
ssh.open(con, connData$server_sas_ip, connData$server_sas_port, connData$server_sas_user, connData$server_sas_pass)
ssh.sendCommand(con, strCMD)
ssh.close(con)
}
}

