% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_base.R
\name{ssh.sendCommand}
\alias{ssh.sendCommand}
\title{ssh.sendCommand}
\usage{
ssh.sendCommand(jShellServer, command, flushwait = FALSE)
}
\examples{
\dontrun{
con = ssh.shell()
ssh.open(con, connData$server_sas_ip, connData$server_sas_port, connData$server_sas_user, connData$server_sas_pass)
ssh.sendCommand(con, strCMD)
ssh.close(con)
}
}

