% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_base.R
\name{ssh.setResponseMultFactorTimeout}
\alias{ssh.setResponseMultFactorTimeout}
\title{ssh.setResponseMultFactorTimeout}
\usage{
ssh.setResponseMultFactorTimeout(jShellServer, responseMultFactorTimeout = 2)
}

