% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_base.R
\name{ssh.shell}
\alias{ssh.shell}
\title{ssh.shell}
\usage{
ssh.shell(jar_path = "", changeBashPromt = TRUE, responseTries = 3,
  responseTimeout = 200, responseMultFactorTries = 2,
  responseMultFactorTimeout = 2, removeSpecialCharactersAtJAVA = TRUE,
  removeSpecialCharactersAtR = TRUE, showConfig = FALSE)
}
\description{
Creates ShellSSH Java Object.
}
\examples{
\dontrun{
con = ssh.shell()
ssh.open(con, connData$server_sas_ip, connData$server_sas_port, connData$server_sas_user, connData$server_sas_pass)
ssh.sendCommand(con, strCMD)
ssh.close(con)
}
}

